Refactor to reuse render method of station browser items
