+++ src/Modules/User/Components/Profile/Profile.tsx
@@ -5,4 +5,3 @@ export { UserDropdown } from './UserDropdown';
 export { Favorite } from './Favorite';
 export { MyStationsBrowser } from './MyStationsBrowser';
 export { RecentStationsBrowser } from './RecentStationsBrowser';
-export { BaseStationBrowser } from './BaseStationBrowser';
