+++ src/Modules/User/Components/Profile/RecentStationsBrowser/RecentStationsBrowser.tsx
@@ -1,11 +1,23 @@
-import { Station } from 'Models';
-import { BaseStationBrowser } from '../BaseStationBrowser';
+import { BaseStationBrowser } from 'BaseComponent/BaseStationBrowser';
+import { StationItem } from 'Models';
+import { UserServices } from 'Services/Http';
 
-export class RecentStationsBrowser extends BaseStationBrowser {
+export class RecentStationsBrowser extends BaseStationBrowser<{}> {
+  public userServices: UserServices;
+
+  constructor(props: {}) {
+    super(props);
+
+    this.userServices = new UserServices();
+  }
+
+  public componentWillMount() {
+    this.getListStation();
+  }
 
   public getListStation() {
     this.userServices.getListMyRecentStation().subscribe(
-      (listStation: Station[]) => {
+      (listStation: StationItem[]) => {
         this.setState({ listStation });
       },
       (err: string) => {
